backend/requirements.txt
aniso8601==9.0.1
Babel==2.10.3
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
mysql-connector-python==8.0.27
mysqlclient==2.0.3
packaging==21.3
passlib==1.7.4
protobuf==4.21.3
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.39
Werkzeug==2.2.0
WTForms==3.0.1
zipp==3.8.1
